Organizing the Front end
